+++ lib/pages/main/views/add_language.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:q_overlay/q_overlay.dart';
 import '../../../helpers/constants.dart';
 import '../controllers/main_controller.dart';
 
@@ -11,10 +12,12 @@ class AddLanguageIcon extends StatelessWidget {
     return IconButton(
         tooltip: 'Add Language',
         onPressed: () async {
-          final value = await Get.dialog<String>(_GetLanguage());
-          if (value == null || value.isEmpty) {
-            return;
-          }
+          final value = await QPanel(
+            width: 200,
+            child: _GetLanguage(),
+            alignment: Alignment.center,
+          ).show<String>();
+          if (value == null || value.isEmpty) return;
           final con = Get.find<MainController>();
           con.locals().languages.add(value);
           con.locals().ensureAllLanguagesExist(con.locals().languages);
@@ -32,23 +35,32 @@ class _GetLanguage extends StatelessWidget {
   final controller = TextEditingController();
   @override
   Widget build(BuildContext context) {
-    return AlertDialog(
-      title: const Text('Add Language'),
-      content: Column(
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
+          const Text('Add Language'),
           TextFormField(controller: controller),
           const Text('Ex. [en] [en_US] [ar] [de]'),
+          const SizedBox(height: 8),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.end,
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              const Spacer(),
+              TextButton(
+                  onPressed: QOverlay.dismissLast, child: const Text('Cancel')),
+              const SizedBox(width: 8),
+              TextButton(
+                  onPressed: () {
+                    QOverlay.dismissLast<String>(result: controller.text);
+                  },
+                  child: const Text('Ok')),
+            ],
+          )
         ],
       ),
-      actions: [
-        ElevatedButton(onPressed: Get.back, child: const Text('Cancel')),
-        ElevatedButton(
-            onPressed: () {
-              Get.back<String>(result: controller.text);
-            },
-            child: const Text('Ok')),
-      ],
     );
   }
 }