+++ lib/pages/main/views/appbar/actions_widget.dart
@@ -4,8 +4,8 @@ import '../../../../widgets/import/import_icon.dart';
 
 import '../../../settings/settings_icon.dart';
 import '../add_language.dart';
-import '../exit_icon.dart';
-import '../save_data_widget.dart';
+import 'exit_icon.dart';
+import 'save_data_widget.dart';
 import 'expand_action.dart';
 
 class ActionsWidget extends StatelessWidget {