+++ lib/widgets/export/controllers/export_controller.dart
@@ -1,8 +1,8 @@
 import 'package:get/get.dart';
-import '../../../models/notification.dart';
 import '../../../pages/main/controllers/main_controller.dart';
 import '../../../services/exporters/easy_localization_exporter_servvice.dart';
 import '../../../services/exporters/getx_exporter_servvice.dart';
+import '../../notification.dart';
 
 class ExportController extends GetxController {
   final RxString path = Get.find<MainController>().appfile.exportPath.obs;
@@ -22,7 +22,7 @@ class ExportController extends GetxController {
     }
 
     mainCon.loading.value = false;
-    mainCon.notification(QNotification.success(message: 'Export Success'));
+    showNotification('Export Success', 'Locals exported successfully');
   }
 }